feat(distribution): Symbolic distribution representation
A subclass of the abstract class Distribution was created to model a generic distribution having an arbitrary functional
formula. All other distributions (Bernoulli, Beta, Gaussian, etc.) are now subclasses of the GenericDistribution class.
Each of the specific distributions provides a formula in symbolic form during initialization. The symbolic form is used
when operations between distributions occur to create a GenericDistribution object. For instance if we multiply a Beta
and a Binomial distribution, the result is a GenericDistribution that has a PDF generated from the multiplication of the
symbolic formulas of Beta and Binomial distributions. That allows for computing PDFs of arbitrary distributions that may
be generated during Bayesian inference.